feat: colocado mensagem de confirmação

+++ src/main/java/net/guilhermejr/sistema/supermercadoservice/config/TarefasAgendadasConfig.java
@@ -6,7 +6,6 @@ import net.guilhermejr.sistema.supermercadoservice.api.request.URLRequest;
 import net.guilhermejr.sistema.supermercadoservice.domain.entity.NFE;
 import net.guilhermejr.sistema.supermercadoservice.domain.repository.NFERepository;
 import net.guilhermejr.sistema.supermercadoservice.service.CompraService;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.scheduling.annotation.EnableScheduling;
 import org.springframework.scheduling.annotation.Scheduled;
@@ -28,39 +27,46 @@ public class TarefasAgendadasConfig {
         log.info("Verificando se existe NFE a ser processada");
         List<NFE> nfeListagemResponses = nfeRepository.findAllByOrderByCriadoAsc();
 
-        nfeListagemResponses.forEach(nfe -> {
+        if (nfeListagemResponses.isEmpty()) {
 
-            URLRequest urlRequest = URLRequest.builder().url(nfe.getUrl()).build();
+            log.info("Nenhuma NFE a ser processada");
 
-            try {
+        } else {
 
-                compraService.inserir(urlRequest, nfe.getUsuario());
+            nfeListagemResponses.forEach(nfe -> {
 
-                log.info("Removido NFE pois foi processada: {}", urlRequest.getUrl());
-                nfeRepository.delete(nfe);
+                URLRequest urlRequest = URLRequest.builder().url(nfe.getUrl()).build();
 
-            } catch (Exception e) {
+                try {
 
-                if (e.getMessage().equals("Compra já realizada")) {
+                    compraService.inserir(urlRequest, nfe.getUsuario());
 
-                    log.info("Removendo NFE pois ela já foi inserida anteriormente: {}", urlRequest.getUrl());
+                    log.info("Removido NFE pois foi processada: {}", urlRequest.getUrl());
                     nfeRepository.delete(nfe);
 
-                } else {
+                } catch (Exception e) {
 
-                    log.error("Erro ao processar NFE: {}", e.getMessage());
-                    Integer tentativas = nfe.getTentativas();
-                    tentativas++;
-                    nfe.setTentativas(tentativas);
-                    nfeRepository.save(nfe);
+                    if (e.getMessage().equals("Compra já realizada")) {
+
+                        log.info("Removendo NFE pois ela já foi inserida anteriormente: {}", urlRequest.getUrl());
+                        nfeRepository.delete(nfe);
+
+                    } else {
+
+                        log.error("Erro ao processar NFE: {}", e.getMessage());
+                        Integer tentativas = nfe.getTentativas();
+                        tentativas++;
+                        nfe.setTentativas(tentativas);
+                        nfeRepository.save(nfe);
+
+                    }
 
                 }
 
-            }
+            });
 
-        });
+        }
 
     }
-
 
 }
